+++ src/modules/asset/Module.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Module for asset files management
+ */
+
+namespace dezero\modules\asset;
+
+class Module extends \dezero\base\Module
+{
+    /**
+     * Initializes the module.
+     *
+     * This method is called after the module is created and initialized with property values
+     * given in configuration. The default implementation will initialize [[controllerNamespace]]
+     * if it is not set.
+     */
+    public function init()
+    {
+        parent::init();
+    }
+}